Properly handle check and set when it is set to 0.
When 'check_and_set' is set to 0, the query param is ommitted completly
which is valid for usecase when we want to update regardles of current
'modify_index'. This however make it impossible to execute requests
with 'check_and_set' set to 0, which is valid use case for setting key
only if it does not already exists.

Signed-off-by: Szymon Gibała <szymongib@gmail.com>