wrap read_key in a ResponseMeta so callers can access the index
when a delete occurs we need this index as the ModifyIndex of all the
ReadKeyResponses will still be the same (and lower). which causes blocking
requests to be "broken" as future read_key request immediatly returns until
either a new key, or a update occuries to a key contained in the
ReadKeyResponse Vec.

per Roblox/rs-consul#30