Switch from license-file to license
According to the cargo docs, the license field is used when you have
a standard license that has an existing SPDX 2.3 license expression.
The license-file field should be used when you are using a non-standard
license type that needs more details.

So why does this matter?  Some automated tools will parse these fields
and use them as an input to license compliance checks.  When using the
license-file field it will go a more complicated path, where an MIT
license should just get a rubber stamp from compliance.

This change has no impact on the actual licensing, it just reflects it
better in a machine parsable way